Ajout liens sur logo et basic menu, finitions sur l'ajout d'un com
