+++ src/components/AppHeader/AppHeader.scss
@@ -72,9 +72,3 @@
       }
 
 }
-
-.account-logo {
-    width: 50px;
-    height: 50px;
-    margin-right: 45px;
-}
