+++ src/styles/_variables.scss
@@ -15,8 +15,8 @@
 // $color-bg: #282c34;
 
 $color-white: #fff;
-$color-primary: hsl(170, 49%, 52%);
-$color-primary-lighter: hsl(170, 49%, 67%);
-$color-primary-darker: hsl(170, 49%, 37%);
+$color-primary: hsl(170, 49%, 52%); // #49c1ad
+$color-primary-lighter: hsl(170, 49%, 67%); // #82d4c6
+$color-primary-darker: hsl(170, 49%, 37%); // #308d7d
 $color-grey: hsl(225, 9%, 70%);
 $color-dark: #141518;
